soom/cjl cast while moving is baseline, removed from wots
